refactor: Migrate theming to Aviato
