+++ apps/docs/package.json
@@ -18,7 +18,6 @@
   },
   "dependencies": {
     "next": "12.0.7",
-    "next-themes": "^0.0.15",
     "react": "17.0.2",
     "react-dom": "17.0.2"
   },
