+++ packages/ui/src/theme/index.ts
@@ -1,0 +1,2 @@
+export * from './theme'
+export * from './provider'
